requirements.txt
tweepy==3.7.0
python-dotenv==0.10.3